envs/autoscale_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-autoscale-env"
version = "0.1.0"
description = "Autoscaling simulator environment packaged for OpenEnv"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.1",
  "fastapi>=0.115.0",
  "pydantic>=2.0.0",
  "uvicorn>=0.24.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
server = "autoscale_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["autoscale_env", "autoscale_env.server"]
package-dir = { "autoscale_env" = ".", "autoscale_env.server" = "server" }
